hitrace: Rename ffi module to api10
